+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/commands/GrabBothClaw.java
@@ -8,7 +8,7 @@ public class GrabBothClaw extends SequentialCommandGroup {
 
     public GrabBothClaw() {
         super(
-                
+
         );
     }
 }
