+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/hardware/drivetrain.java
@@ -7,8 +7,6 @@ import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.IMU;
 
-import org.firstinspires.ftc.teamcode.common.SampleMecanumDrive;
-
 import java.util.Timer;
 import java.util.TimerTask;
