+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/testing/AutomaticFeedforwardTuner.java
@@ -15,7 +15,7 @@ import com.qualcomm.robotcore.util.RobotLog;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.internal.system.Misc;
-import org.firstinspires.ftc.teamcode.common.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcode.common.hardware.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.common.util.LoggingUtil;
 import org.firstinspires.ftc.teamcode.common.util.RegressionUtil;
