+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/testing/MotorDirectionDebugger.java
@@ -7,7 +7,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.common.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcode.common.hardware.SampleMecanumDrive;
 
 /**
  * This is a simple teleop routine for debugging your motor configuration.
