+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/testing/StraightTest.java
@@ -10,7 +10,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.common.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcode.common.hardware.SampleMecanumDrive;
 
 /*
  * This is a simple routine to test translational drive capabilities.
