+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/testing/TrackingWheelForwardOffsetTuner.java
@@ -13,8 +13,8 @@ import com.qualcomm.robotcore.util.RobotLog;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.internal.system.Misc;
-import org.firstinspires.ftc.teamcode.common.SampleMecanumDrive;
-import org.firstinspires.ftc.teamcode.common.StandardTrackingWheelLocalizer;
+import org.firstinspires.ftc.teamcode.common.hardware.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcode.common.hardware.StandardTrackingWheelLocalizer;
 
 /**
  * This routine determines the effective forward offset for the lateral tracking wheel.
